Moved work models into their own app
