+++ tbx/core/templates/torchbox/tags/work_list_item.html
@@ -1,1 +1,1 @@
-{% include "torchbox/includes/work_item.html" with page=work title=work.title descriptive_title=work.descriptive_title image=work.homepage_image %}
+{% include "work/includes/work_item.html" with page=work title=work.title descriptive_title=work.descriptive_title image=work.homepage_image %}
