+++ tbx/services/templates/blocks/services/case_study_block.html
@@ -11,7 +11,7 @@
             <ul class="work-list">
                 {% for case_study in value.case_studies %}
                 {% with case_study.page.specific as page %}
-                  {% include "torchbox/includes/work_item.html" with page=page title=case_study.title|default:page.title descriptive_title=case_study.descriptive_title|default:page.descriptive_title image=case_study.image|default:page.homepage_image %}
+                  {% include "work/includes/work_item.html" with page=page title=case_study.title|default:page.title descriptive_title=case_study.descriptive_title|default:page.descriptive_title image=case_study.image|default:page.homepage_image %}
                 {% endwith %}
                 {% endfor %}
             </ul>
